Move ARMA to new module misc.noise
