+++ PyDynamic/uncertainty/propagate_MonteCarlo.py
@@ -29,6 +29,7 @@ from scipy.signal import lfilter
 
 from ..misc.filterstuff import isstable
 from ..misc.tools import progress_bar
+from ..misc.noise import ARMA
 
 __all__ = ["MC", "SMC", "UMC", "UMC_generic"]
 
@@ -556,57 +557,6 @@ def _UMCevaluate(th, nbb, x, Delta, phi, theta, sigma, blow, alow):
     d = Delta * (2 * np.random.random_sample(size=x.size) - 1 )   # uniform distribution [-Delta, Delta]
 
     return lfilter(bb, aa, xlow) + d
-
-
-# move this function to ..misc.noise.ARMA
-def ARMA(length, phi = 0.0, theta = 0.0, std = 1.0):
-    """
-    Generate time-series of a predefined ARMA-process based on this equation:
-    :math:`\sum_{j=1}^{\min(p,n-1)} \phi_j \epsilon[n-j] + \sum_{j=1}^{\min(q,n-1)} \\theta_j w[n-j]`
-    where w is white gaussian noise. Equation and algorithm taken from [Eichst2012]_ .
-
-    Parameters
-    ----------
-    length: int
-        how long the drawn sample will be
-    phi: float, list or numpy.ndarray, shape (p, )
-        AR-coefficients
-    theta: float, list or numpy.ndarray,
-        MA-coefficients
-    std: float
-        std of the gaussian white noise that is feeded into the ARMA-model
-
-    References
-    ----------
-        * Eichstädt, Link, Harris and Elster [Eichst2012]_
-    """
-
-    # convert to numpy.ndarray
-    if isinstance(phi, float):
-        phi = np.array([phi])
-    elif isinstance(phi, list):
-        phi = np.array(phi)
-
-    if isinstance(theta, float):
-        theta = np.array([theta])
-    elif isinstance(theta, list):
-        theta = np.array(theta)
-
-    # initialize e, w
-    w = np.random.normal(loc = 0, scale = std, size = length)
-    e = np.zeros_like(w)
-
-    # define shortcuts
-    p = len(phi)
-    q = len(theta)
-
-    # iterate series over time
-    for n, wn in enumerate(w):
-        min_pn = min(p, n)
-        min_qn = min(q, n)
-        e[n] = np.sum(phi[:min_pn].dot(e[n-min_pn:n])) + np.sum(theta[:min_qn].dot(w[n-min_qn:n])) + wn
-
-    return e
 
 
 def UMC_generic(draw_samples, evaluate, runs = 100, blocksize = 8, runs_init = 10, nbins = 100, return_simulations = False, n_cpu = multiprocessing.cpu_count()):
